+++ src/main/kotlin/io/kjson/JSONCoPipeline.kt
@@ -45,6 +45,7 @@ import net.pwall.pipeline.CoAcceptor
  * emits JSON values to a non-blocking downstream receiver as they are completed.
  *
  * @author  Peter Wall
+ * @param   <R>     the pipeline result type
  */
 class JSONCoPipeline<R> internal constructor(
     downstream: CoAcceptor<JSONValue?, R>,